backend/requirements.txt
asgiref==3.8.1
sqlparse==0.5.2
gunicorn==23.0.0
Django==5.2.8
django-filter==25.2
djangorestframework==3.16.1
drf-spectacular==0.29.0
Faker==38.2.0
psycopg==3.2.12
psycopg-binary==3.2.12
python-dotenv==1.2.1